Extract translation-related functions into separate module.
